Removes ValueOwned dependency on eval Value
Removes LegacySpec
