+++ partiql-eval/src/engine/value/mod.rs
@@ -2,6 +2,7 @@
 mod register_reader;
 mod shape;
 mod value_owned;
+mod value_view;
 
 // Internal types
 mod internal;
@@ -10,7 +11,7 @@ mod internal;
 pub use register_reader::RegisterReader;
 pub use shape::{FieldName, FieldShape, PhysicalType, RowShape, Shape};
 pub(crate) use value_owned::ValueOwned;
-pub use value_owned::{ValueType, ValueView};
+pub use value_view::{ValueType, ValueView};
 
 // Internal exports for use within the engine
 pub(crate) use internal::{value_get_field_ref, TupleField, TupleRef, ValueRef};
